requirements.txt
aiohttp==3.9.5
discord.py==2.4.0
